requirements.txt
click==8.1.3
Flask==2.3.2
gunicorn==20.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
flask_sqlalchemy==3.0.5
psycopg2==2.9.6
